wayland: share xkb with x11 backend
