+++ druid-shell/Cargo.toml
@@ -20,8 +20,8 @@ x11 = ["x11rb", "nix", "cairo-sys-rs", "bindgen", "pkg-config"]
 # Implement HasRawWindowHandle for WindowHandle
 raw-win-handle = ["raw-window-handle"]
 # **WARNING** not ready for the prime time. Many things don't work yet.
-wayland = ["wayland-client", "wayland-protocols/client", "wayland-protocols/unstable_protocols", 
-"nix", "cairo-sys-rs", "rand", "xkbcommon-sys", "calloop", "wayland-cursor", "log", "im"]
+wayland = ["wayland-client", "wayland-protocols/client", "wayland-protocols/unstable_protocols",
+"nix", "cairo-sys-rs", "rand", "calloop", "wayland-cursor", "log", "im", "bindgen", "pkg-config"]
 
 # passing on all the image features. AVIF is not supported because it does not
 # support decoding, and that's all we use `Image` for.
@@ -94,7 +94,6 @@ wayland-client = { version = "0.28.2", optional = true, features = ["dlopen"] }
 wayland-protocols = { version = "0.28.2", optional = true }
 wayland-cursor = { version = "0.28.3", optional = true }
 rand = { version = "0.8.0", optional = true }
-xkbcommon-sys = { version = "0.7.4", optional = true }
 calloop = { version = "0.7.1", optional = true }
 log = { version = "0.4.14", optional = true }
 im = { version = "15.0.0", optional = true }
