+++ druid-shell/src/backend/shared/mod.rs
@@ -24,5 +24,6 @@ cfg_if::cfg_if! {
     if #[cfg(all(target_os = "linux", any(feature = "x11", feature = "wayland")))] {
         mod timer;
         pub(crate) use timer::*;
+        pub(crate) mod xkb;
     }
 }
