+++ druid-shell/src/backend/shared/xkb/mod.rs
@@ -14,19 +14,23 @@
 
 //! A minimal wrapper around Xkb for our use.
 
-use super::keycodes;
-use super::xkbcommon_sys::*;
+mod keycodes;
+mod xkbcommon_sys;
 use crate::{
     backend::shared::{code_to_location, hardware_keycode_to_code},
     KeyEvent, KeyState, Modifiers,
 };
 use keyboard_types::{Code, Key};
 use std::convert::TryFrom;
-use std::os::raw::{c_char, c_int};
+use std::os::raw::c_char;
 use std::ptr;
+use xkbcommon_sys::*;
+
+#[cfg(feature = "x11")]
 use x11rb::xcb_ffi::XCBConnection;
 
-pub struct DeviceId(c_int);
+#[cfg(feature = "x11")]
+pub struct DeviceId(std::os::raw::c_int);
 
 /// A global xkb context object.
 ///
@@ -42,6 +46,7 @@ impl Context {
         unsafe { Self(xkb_context_new(XKB_CONTEXT_NO_FLAGS)) }
     }
 
+    #[cfg(feature = "x11")]
     pub fn core_keyboard_device_id(&self, conn: &XCBConnection) -> Option<DeviceId> {
         let id = unsafe {
             xkb_x11_get_core_keyboard_device_id(
@@ -55,6 +60,7 @@ impl Context {
         }
     }
 
+    #[cfg(feature = "x11")]
     pub fn keymap_from_device(&self, conn: &XCBConnection, device: DeviceId) -> Option<Keymap> {
         let key_map = unsafe {
             xkb_x11_keymap_new_from_device(
@@ -68,6 +74,30 @@ impl Context {
             return None;
         }
         Some(Keymap(key_map))
+    }
+
+    /// Create a keymap from some given data.
+    ///
+    /// Uses `xkb_keymap_new_from_buffer` under the hood.
+    #[cfg(feature = "wayland")]
+    pub fn keymap_from_slice(&self, buffer: &[u8]) -> Keymap {
+        // TODO we hope that the keymap doesn't borrow the underlying data. If it does' we need to
+        // use Rc. We'll find out soon enough if we get a segfault.
+        // TODO we hope that the keymap inc's the reference count of the context.
+        assert!(
+            buffer.iter().copied().any(|byte| byte == 0),
+            "`keymap_from_slice` expects a null-terminated string"
+        );
+        unsafe {
+            let keymap = xkb_keymap_new_from_string(
+                self.0,
+                buffer.as_ptr() as *const i8,
+                XKB_KEYMAP_FORMAT_TEXT_V1,
+                XKB_KEYMAP_COMPILE_NO_FLAGS,
+            );
+            assert!(!keymap.is_null());
+            Keymap(keymap)
+        }
     }
 
     /// Set the log level using `tracing` levels.
