+++ druid-shell/src/backend/wayland/application.rs
@@ -16,10 +16,12 @@
 
 use super::{
     clipboard::Clipboard, error::Error, events::WaylandSource, keyboard, pointers, surfaces,
-    window::WindowHandle,
+    surfaces::buffers::Mmap, window::WindowHandle,
 };
 
-use crate::{application::AppHandler, backend, kurbo, mouse, TimerToken};
+use crate::{
+    backend, backend::shared::xkb, keyboard_types::KeyState, kurbo, mouse, AppHandler, TimerToken,
+};
 
 use calloop;
