--- druid-shell/src/backend/wayland/keycodes.rs
@@ -1,29 +1,0 @@
-// Copyright 2019 The Druid Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use crate::keyboard_types::{Key, Location};
-
-pub type RawKey = ();
-
-pub fn raw_key_to_key(raw: RawKey) -> Option<Key> {
-    todo!()
-}
-
-pub fn raw_key_to_location(raw: RawKey) -> Location {
-    todo!()
-}
-
-pub fn key_to_raw_key(src: &Key) -> Option<RawKey> {
-    todo!()
-}
