+++ druid-shell/src/backend/wayland/mod.rs
@@ -18,7 +18,7 @@ pub mod application;
 pub mod clipboard;
 pub mod dialog;
 pub mod error;
-pub mod events;
+mod events;
 pub mod keyboard;
 pub mod menu;
 pub mod pointers;
@@ -26,7 +26,6 @@ pub mod screen;
 pub mod surfaces;
 pub mod util;
 pub mod window;
-pub mod xkb;
 
 /// Little enum to make it clearer what some return values mean.
 #[derive(Copy, Clone)]
